backend/requirements.txt
# Fingerprint Liveness Risk Analysis System
# Python 3.10+ required

# =============================================================================
# Web Framework
# =============================================================================
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6         # For file uploads

# =============================================================================
# Machine Learning / Deep Learning
# =============================================================================
tensorflow>=2.16.0              # CNN model training and inference
numpy>=1.24.0                   # Numerical operations
scipy>=1.11.0                   # Required by Keras ImageDataGenerator

# =============================================================================
# Image Processing
# =============================================================================
opencv-python>=4.8.0            # Image decoding and preprocessing
Pillow>=10.0.0                  # Image generation for demo dataset

# =============================================================================
# Data Validation
# =============================================================================
pydantic>=2.0.0                 # Request/response validation
